+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This changelog also contains important changes in dependencies.
 - (usvg) Style attributes resolving on the root `svg` element.
 - (usvg) `SmoothCurveTo` and `SmoothQuadratic` conversion.
 - (usvg) `symbol` resolving.
+- (cairo-backend) Font ascent calculation.
 - (qt-backend) Stroking of LineTo specified as CurveTo.
 - (svgdom) `stroke-miterlimit` attribute parsing.
 - (svgdom) `length` and `number` attribute types parsing.